add dhcpv4.GetOption and dhcpv4.GetSingleOption methods
GetOption and GetOneOption are convience methods for getting a DHCPv4
option from it's Option Code.

GetOption returns a list of options that match since the DHCP RFC allows
for an option to be present multiple times with it's values appened
together.  (Note: I am not sure i've ever seen this done and it's not clear if
how option parsing would work for more complex values.  Is appeneding
done at the byte level? )

GetOneOption will return the first found option that maches a code.