add dhcpv4.GetOption and dhcpv4.GetSingleOption methods (#63)
GetOption and GetOneOption are convenience methods for getting a DHCPv4
option from its Option Code.

GetOption returns a list of options that match since the DHCP RFC allows
for an option to be present multiple times with its values appended
together.  (Note: I am not sure I've ever seen this done and it's not clear if
how option parsing would work for more complex values.  Is appending
done at the byte level? )

GetOneOption will return the first found option that matches a code.