Pass through unknown DUIDs
I’m building on a program which works with user-supplied DUIDs. Before this
change, the code would panic when users provide a DUID which is not (yet)
implemented.

In addition to this being a better failure mode, the behavior is also mandated
by the DHCPv6 RFC: https://tools.ietf.org/html/rfc3315#section-9 states:

    Clients and servers MUST treat DUIDs as opaque values and MUST only
    compare DUIDs for equality.