Allow Unknown OperState of the link interface (#254)
* Allow Unknowo OperState of the link interface

We hig a bug when Netconf library was failing to bring interface up
despite the fact that it was actually up. It turned out that it's oper
state was not set to UP, what is expected by the library.

According to kernel documentation it is ok proceed if interface state
is Up or Unknown:

```
  Interface is in RFC2863 operational state UP or UNKNOWN. This is for
  backward compatibility, routing daemons, dhcp clients can use this
  flag to determine whether they should use the interface.
```

Also, resaon why operational state may remain Unknown:

```
IF_OPER_UNKNOWN (0):
 Interface is in unknown state, neither driver nor userspace has set
 operational state. Interface must be considered for user data as
 setting operational state has not been implemented in every driver.
```

I modified our code to try DHCP transaction even if `IfUp` failed, but
the OperState was equal to Unknown - it worked perfectly.

* Skip rt7 test also with go 1.10 and 1.11

As per request from @pmazzini.