switch from u-root/pkg/rand to crypto/rand
As per its documentation, u-root’s rand package implements cancelable reads,
which is not something that this DHCP client uses.

The crypto/rand package has the advantage of using getrandom(2) on platforms
where that is possible, and otherwise behaves exactly like u-root/pkg/rand.

Notably, this commit fixes the build when running with GO111MODULES=on, so it’s
related to issue #123.

Signed-off-by: Michael Stapelberg <stapelberg@google.com>