Increase DHCPv4 IP TTL from 30 to 64
RFC 1700 recommends a value of 64 for the default IP time to live (TTL)
parameter. This is only necessary for the V4 client because it uses
the net.PacketConn, while the IPv6 implementation uses a net.UDPConn instead.

From https://tools.ietf.org/html/rfc1700:

IP TIME TO LIVE PARAMETER

The current recommended default time to live (TTL) for the
Internet Protocol (IP) [45,105] is 64.

Signed-off-by: Ross Hanson <rosshanson@google.com>