netconf: apply v6 addresses as /128
https://bugs.debian.org/cgi-bin/bugreport.cgi?bug=684009

"Note that the dhcpv6 protocol doesn't have an option for a netmask. So
it is always /128 and routing is left to icmpv6 router advertisements."

RFC 5942 is a good read here as well:

   An address could be acquired through the DHCPv6 identity association
   for non- temporary addresses (IA_NA) option from [RFC3315] (which
   does not include a prefix length), or through manual configuration
   (if no prefix length is specified).  The host incorrectly assumes an
   invented prefix is on-link.  This invented prefix typically is a /64
   that was written by the developer of the operating system network
   module API to any IPv6 application as a "default" prefix length when
   a length isn't specified.

As DHCP developers, we *HAVE* to assume that no prefix is on-link. The
correct way to do that is to specify the netmask as /128.

The kernel will RA/RS their way around to figure out what prefixes are
indeed on-link.

Signed-off-by: Chris Koch <chrisko@google.com>