Pad packets after the 'end' option to be more conformant
RFC 2131 gives two examples of padding in section 4.1. For both the
examples, the 'end' option preceeds the padding with the 'pad' option.
These examples relate to overloading pre-defined fields for options,
not padding the packet for a minimal size.

Older versions of the U-Boot firmware support DHCP, but do not expect
the 'pad' option. The firmware logs the following warning for each
occurence:
    *** Unhandled DHCP Option in OFFER/ACK: 0
This warning disappears when the 'end' option preceeds the 'pad'
option.

Signed-off-by: Marcel Moolenaar <mmoolena@amazon.com>
Signed-off-by: Hariharakumar Narasimhakumar <hhkumar@fb.com>