Add WithContext versions of functions that use contexts.
These new functions are equivalent, but allow callers to pass a context down from main (or equivalent).  This is necessary to comply with certain style guides, for example, https://google.github.io/styleguide/go/decisions#contexts.

We considered replacing dhcpv4.GenerateTransactionID with rand.Read(xid[:]) based on the godoc saying that Read call "never returns an error", but this package is used in embedded situations where it can fail.

Signed-off-by: Mark Suter <suter@zwitterion.org>